fix(mobile): 3 QA bugs — Director canvas reachable, Gallery toolbar, Profile date
- Director: limit left panel to 60vh on mobile so canvas area is scrollable
- Gallery: show icon instead of "Editar en Lote" text on small screens
- Profile: use abbreviated month format (mar 2026) to prevent stat card wrapping

+++ components/ProfilePage.tsx
@@ -118,7 +118,7 @@ const ProfilePage: React.FC = () => {
     .join('');
 
   const memberSince = profile?.createdAt
-    ? new Date(profile.createdAt).toLocaleDateString('es-MX', { month: 'long', year: 'numeric' })
+    ? new Date(profile.createdAt).toLocaleDateString('es-MX', { month: 'short', year: 'numeric' })
     : '—';
 
   if (isLoading) {
